+++ drivers/video/tegra/dc/dc_config.h
@@ -167,7 +167,7 @@ enum tegra_dc_feature_option {
 struct tegra_dc_feature_entry {
 	u32 window_index;
 	u32 option;
-	long arg[ENTRY_SIZE];
+	u32 arg[ENTRY_SIZE];
 };
 
 struct tegra_dc_feature {
@@ -182,7 +182,7 @@ int tegra_dc_feature_has_interlace(struct tegra_dc *dc, int win_idx);
 int tegra_dc_feature_has_filter(struct tegra_dc *dc, int win_idx, int operation);
 int tegra_dc_feature_is_gen2_blender(struct tegra_dc *dc, int win_idx);
 
-long *tegra_dc_parse_feature(struct tegra_dc *dc, int win_idx, int operation);
+u32 *tegra_dc_parse_feature(struct tegra_dc *dc, int win_idx, int operation);
 void tegra_dc_feature_register(struct tegra_dc *dc);
 
 static inline bool win_use_v_filter(struct tegra_dc *dc,
